Chunk worker output to 1024 characters
When we output too much data on the zpty fd it can become corrupt in
some cases, this commit protects against this by only printing 1024
bytes at a time.

Further, when notifying via kill signas, we notify the parent process
after every chunk so that the fd can be emptied.

In my testing, 1024 bytes seems to be the maximum safe limit that can be
used. Perhaps evidence to support this conclusion could be found in the
zsh code base.